+++ src/main/java/org/realityforge/replicant/server/ee/AbstractReplicationInterceptor.java
@@ -1,16 +1,7 @@
 package org.realityforge.replicant.server.ee;
 
-import java.util.Collection;
-import java.util.Collections;
-import javax.annotation.Resource;
 import javax.interceptor.AroundInvoke;
 import javax.interceptor.InvocationContext;
-import javax.persistence.EntityManager;
-import javax.transaction.TransactionSynchronizationRegistry;
-import org.realityforge.replicant.server.ChangeSet;
-import org.realityforge.replicant.server.EntityMessage;
-import org.realityforge.replicant.server.EntityMessageEndpoint;
-import org.realityforge.replicant.server.EntityMessageSet;
 import org.realityforge.replicant.shared.transport.ReplicantContext;
 
 /**
@@ -19,62 +10,37 @@ import org.realityforge.replicant.shared.transport.ReplicantContext;
  */
 public abstract class AbstractReplicationInterceptor
 {
-  /**
-   * The key used to access the registry to get the current call depth.
-   */
-  private static final String REPLICATION_TX_DEPTH = "ReplicationTxDepth";
-
-  @Resource
-  private TransactionSynchronizationRegistry _registry;
-
   @AroundInvoke
   public Object businessIntercept( final InvocationContext context )
     throws Exception
   {
-    final Integer depth = (Integer) _registry.getResource( REPLICATION_TX_DEPTH );
-    if ( null == depth )
+    final int depth = getRequestManager().getReplicationCallDepth();
+    if ( 0 == depth )
     {
       final String sessionID = (String) ReplicantContextHolder.get( ReplicantContext.SESSION_ID_KEY );
       if ( null != sessionID )
       {
-        _registry.putResource( ReplicantContext.SESSION_ID_KEY, sessionID );
         ReplicantContextHolder.remove( ReplicantContext.SESSION_ID_KEY );
       }
       final String requestID = (String) ReplicantContextHolder.get( ReplicantContext.REQUEST_ID_KEY );
       if ( null != requestID )
       {
-        _registry.putResource( ReplicantContext.REQUEST_ID_KEY, requestID );
         ReplicantContextHolder.remove( ReplicantContext.REQUEST_ID_KEY );
       }
+
+      getRequestManager().startReplication( sessionID, requestID );
     }
-    final String sessionID = (String) _registry.getResource( ReplicantContext.SESSION_ID_KEY );
-    final String requestID = (String) _registry.getResource( ReplicantContext.REQUEST_ID_KEY );
-    _registry.putResource( REPLICATION_TX_DEPTH, ( null == depth ? 1 : depth + 1 ) );
+    getRequestManager().setReplicationCallDepth( depth + 1 );
     try
     {
       return context.proceed();
     }
     finally
     {
-      _registry.putResource( REPLICATION_TX_DEPTH, depth );
-      if ( null == depth )
+      getRequestManager().setReplicationCallDepth( depth );
+      if ( 0 == depth )
       {
-        boolean requestComplete = true;
-        if ( getEntityManager().isOpen() && !_registry.getRollbackOnly() )
-        {
-          getEntityManager().flush();
-          final EntityMessageSet messageSet = EntityMessageCacheUtil.removeEntityMessageSet( _registry );
-          final ChangeSet changeSet = EntityMessageCacheUtil.removeSessionChanges( _registry );
-          if ( null != messageSet || null != changeSet )
-          {
-            final Collection<EntityMessage> messages =
-              null == messageSet ? Collections.<EntityMessage>emptySet() : messageSet.getEntityMessages();
-            if ( null != changeSet || messages.size() > 0 )
-            {
-              requestComplete = !getEndpoint().saveEntityMessages( sessionID, requestID, messages, changeSet );
-            }
-          }
-        }
+        final boolean requestComplete = getRequestManager().completeReplication();
         if ( !ReplicantContextHolder.contains( ReplicantContext.REQUEST_COMPLETE_KEY ) )
         {
           ReplicantContextHolder.put( ReplicantContext.REQUEST_COMPLETE_KEY, requestComplete ? "1" : "0" );
@@ -83,7 +49,5 @@ public abstract class AbstractReplicationInterceptor
     }
   }
 
-  protected abstract EntityManager getEntityManager();
-
-  protected abstract EntityMessageEndpoint getEndpoint();
+  protected abstract ReplicationRequestManager getRequestManager();
 }